refactor: Use a `core` folder
